fix(time) : fix time error

+++ philosophers/headers/philosophers.h
@@ -6,7 +6,7 @@
 /*   By: hyojocho <hyojocho@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/14 15:58:04 by hyojocho          #+#    #+#             */
-/*   Updated: 2023/06/24 15:00:55 by hyojocho         ###   ########.fr       */
+/*   Updated: 2023/06/28 12:28:40 by hyojocho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@
 
 // get_args
 int				get_args(int argc, char **argv, t_args_info *args_info);
-unsigned int	get_time(void);
+long long		get_time(void);
 // init_resources
 int				init_resources(t_resources *resources, t_args_info *args_info);
 int				init_time(t_args_info *args_info);
